Added lights and improved the drone
Drone laser gun now separate from drone
Drone now has a spotlight
The drone's look rotation doesn't snap around anymore, smooth rotation only!
