+++ Assets/Enemies/Scripts/FlyingEnemy.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
+using Unity.VisualScripting;
 using UnityEngine;
 using Random = UnityEngine.Random;
 
@@ -12,46 +13,51 @@ public class FlyingEnemy : Enemy
     [SerializeField] protected float speed;
     [SerializeField] protected float accelerationTime;
     [SerializeField] protected float slowdownDistance;
+    [SerializeField] protected float lookSpeed;
     
+    [SerializeField] protected LaserGun laserGun;
+
     protected Vector3 currentVelocity;
     protected float locationTolerance;
     
     protected void Update()
     {
         locationTolerance = range[0];
-
+        // If the enemy can currently see the player
         if (CanSeePlayer(out RaycastHit hit))
         {
             // Debug.Log(hit);
+            var rotation = Quaternion.RotateTowards(
+                laserGun.transform.rotation,
+                Quaternion.LookRotation(Player.Instance.transform.position + Vector3.up - transform.position),
+                lookSpeed);
+            laserGun.transform.rotation = rotation; 
             targetLocation = hit.transform.position;
-            MoveTo(targetLocation);
-            if ((hit.transform.position - transform.position).magnitude < range[1]) 
+            if ((hit.transform.position - transform.position).magnitude < range[1])
                 Attack();
-            return;
-        } 
-        Idle();
-        // Make raycast towards player
-        /*Vector3 direction = Player.Instance.transform.position + new Vector3(0, 1, 0) - transform.position;
-        Ray ray = new Ray(transform.position, direction);
-        if (Physics.Raycast(ray, out RaycastHit hit, viewDistance, visible))
+            MoveTo(targetLocation);
+
+        }
+        // Else the enemy cannot see the player
+        else
         {
-            // If enemy can see player
-            if (hit.transform.gameObject.CompareTag("Player"))
+            Ray ray = new Ray(targetLocation, Vector3.down);
+            if (Physics.Raycast(ray, out RaycastHit groundHit, 512, visible))
             {
-                targetLocation = hit.transform.position;
-                MoveTo(targetLocation);
-                if ((hit.transform.position - transform.position).magnitude < range[1]) 
-                    Attack();
+                var rotation = Quaternion.RotateTowards(
+                    laserGun.transform.rotation,
+                    Quaternion.LookRotation(groundHit.point - transform.position),
+                    lookSpeed);
+                laserGun.transform.rotation = rotation; 
+                // laserGun.transform.LookAt(groundHit.point);
             }
-            // Idle if unable to see player
-            else
-                Idle();
-        }*/
+            Idle();
+        }        
+        
     }
 
     protected virtual void Attack()
     { 
-        transform.LookAt(Player.Instance.transform.position + Vector3.up);
     }
 
     protected void Idle()
@@ -65,19 +71,11 @@ public class FlyingEnemy : Enemy
                 ref currentVelocity, accelerationTime / 2);
             locationTolerance += Time.deltaTime * (locationTolerance + 0.1f);
         }
-
-        /*else if (enemyRigidbody.velocity.magnitude > 0.1f)
-        {
-            enemyRigidbody.velocity = Vector3.SmoothDamp(enemyRigidbody.velocity, Vector3.zero, 
-                ref currentVelocity, accelerationTime / 2);
-        }*/
         else
         {
             targetLocation = GenerateRandomTarget();
             locationTolerance = 1;
         }
-            
-        // Debug.Log("Idle");
     }
 
     protected override Vector3 GenerateRandomTarget()
@@ -94,7 +92,6 @@ public class FlyingEnemy : Enemy
 
     protected void MoveTo(Vector3 target)
     {
-        // Debug.Log("Distance: " + (transform.position - targetLocation).magnitude + "Target: " + target);
         target.y += hoverHeight[0];
         var delta = target - transform.position;
         var direction = delta.normalized;
