+++ Assets/Enemies/Scripts/LaserGun.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7bc5169499c5f934583e76dc19261aef
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
